Users\Files\delete

+++ www/fns/Files/delete.php
@@ -10,14 +10,6 @@ function delete ($mysqli, $id_users, $id) {
     include_once __DIR__.'/filePath.php';
     $filePath = filePath($id_users, $id);
 
-    if (is_file($filePath)) {
-
-        $file_size = filesize($filePath);
-        unlink($filePath);
-
-        include_once __DIR__.'/../Users/addStorageUsed.php';
-        \Users\addStorageUsed($mysqli, $id_users, -$file_size);
-
-    }
+    if (is_file($filePath)) unlink($filePath);
 
 }
